feat: Add publish and unpublish buttons to insight form
