+++ src/locales/en/translation.json
@@ -780,10 +780,12 @@
   "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_CREATE": "Create insight",
   "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_DELETE": "Delete",
   "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_EDIT": "Edit",
+  "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_PUBLISH": "Publish",
+  "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_UNPUBLISH": "Unpublish",
   "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_UNDO": "Cancel",
   "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_UPDATE": "Update insight",
   "__INSIGHTS_PAGE_INSIGHT_FORM_FIELD_DESCRIPTION_LABEL": "Description",
-  "__INSIGHTS_PAGE_INSIGHT_FORM_FIELD_DESCRIPTION_PLACEHOLDER": "Esempio di descrizione:&#10;La maggior parte dei partecipanti (9) ha preferito l'opzione C principalmente perché offre una doppia possibilità di navigazione......",
+  "__INSIGHTS_PAGE_INSIGHT_FORM_FIELD_DESCRIPTION_PLACEHOLDER": "Esempio di descrizione: La maggior parte dei partecipanti (9) ha preferito l'opzione C principalmente perché offre una doppia possibilità di navigazione......",
   "__INSIGHTS_PAGE_INSIGHT_FORM_FIELD_OBSERVATIONS_ERROR": "At least one observation is required",
   "__INSIGHTS_PAGE_INSIGHT_FORM_FIELD_OBSERVATIONS_LABEL": "Observations",
   "__INSIGHTS_PAGE_INSIGHT_FORM_FIELD_OBSERVATIONS_SELECTED": "You selected:",
