+++ src/locales/it/translation.json
@@ -809,6 +809,8 @@
   "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_CREATE": "",
   "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_DELETE": "",
   "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_EDIT": "",
+  "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_PUBLISH": "",
+  "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_UNPUBLISH": "",
   "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_UNDO": "",
   "__INSIGHTS_PAGE_INSIGHT_FORM_BUTTON_UPDATE": "",
   "__INSIGHTS_PAGE_INSIGHT_FORM_FIELD_DESCRIPTION_LABEL": "",
